refactor : vendingMachine -> CoinRepository, ItemRepository 중복 코드 제거
