+++ src/main/java/vendingmachine/service/VendingMachineService.java
@@ -3,32 +3,53 @@ package vendingmachine.service;
 import java.util.Arrays;
 import java.util.List;
 
+import camp.nextstep.edu.missionutils.Randoms;
+import vendingmachine.domain.Coin;
 import vendingmachine.domain.Item;
-import vendingmachine.repository.VendingMachine;
+import vendingmachine.repository.CoinRepository;
+import vendingmachine.repository.ItemRepository;
 import vendingmachine.util.Symbol;
 
 public class VendingMachineService {
+	private static final List<Integer> coin_kind = Arrays.asList(Coin.COIN_10.getAmount(),
+		Coin.COIN_50.getAmount(),
+		Coin.COIN_100.getAmount(),
+		Coin.COIN_500.getAmount());
 	private static final int ZERO = 0;
 
-	private final VendingMachine vendingMachine;
+	private final CoinRepository coinRepository;
+	private final ItemRepository itemRepository;
 
 	public VendingMachineService() {
-		this.vendingMachine = new VendingMachine();
+		this.coinRepository = new CoinRepository(coin_kind);
+		this.itemRepository = new ItemRepository();
 	}
 
 	public void changeMoneyToCoin(int money) {
-		vendingMachine.changeToCoin(money);
+		while (money > ZERO) {
+			int coin = Randoms.pickNumberInList(coin_kind);
+			if (canChange(money, coin)) {
+				coinRepository.addCoin(Coin.fromMoney(coin));
+				money -= coin;
+			}
+		}
+	}
 
+	private boolean canChange(int money, int coin) {
+		if (money - coin >= ZERO) {
+			return true;
+		}
+		return false;
 	}
 
 	public String getMachineSmallChange() {
-		return vendingMachine.getCurrentMachineCoin();
+		return coinRepository.getCurrentMachineCoin();
 	}
 
 	public void saveItem(String itemInfo) {
 		List<String> listOfInfo = Arrays.asList(decodeInput(itemInfo));
 		listOfInfo.stream().map(s -> s.split(","))
-			.forEach(s -> vendingMachine.addItem(s[0], new Item(s[0], Integer.parseInt(s[1]), Integer.parseInt(s[2]))));
+			.forEach(s -> itemRepository.addItem(s[0], new Item(s[0], Integer.parseInt(s[1]), Integer.parseInt(s[2]))));
 	}
 
 	private String[] decodeInput(String itemInfo) {
@@ -38,21 +59,21 @@ public class VendingMachineService {
 	}
 
 	public boolean canBuyAnything(int payMoney) {
-		if (payMoney < vendingMachine.getMinItemPrice() || vendingMachine.isAllItemSoldOut()) {
+		if (payMoney < itemRepository.getMinItemPrice() || itemRepository.isAllItemSoldOut()) {
 			return false;
 		}
 		return true;
 	}
 
 	public int buyItem(int payMoney, String itemName) {
-		if (!vendingMachine.isItemSoldOut(itemName)) {
-			return payMoney - vendingMachine.buyItemAndGetPrice(itemName);
+		if (!itemRepository.isItemSoldOut(itemName)) {
+			return payMoney - itemRepository.buyItemAndGetPrice(itemName);
 		}
 		return payMoney;
 	}
 
 	public String calculateSmallChange(int remainMoney) {
-		return vendingMachine.subtractCoins(remainMoney);
+		return coinRepository.subtractCoins(remainMoney);
 	}
 
 }
